pyproject.toml
[tool.poetry]
name = "cw-7"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.2"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
djangorestframework-simplejwt = "^5.3.1"
celery = "^5.4.0"
eventlet = "^0.36.1"
redis = "^5.0.8"
django-celery-beat = "^2.6.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
